requirements.txt
flask==3.0.0
openpyxl==3.1.2
supabase==2.3.4
python-dotenv==1.0.0
